replace problemNumber to problemId

+++ server/src/main/java/com/example/server/student/problem/ProblemRecommender.java
@@ -3,5 +3,5 @@ package com.example.server.student.problem;
 import java.util.List;
 
 public interface ProblemRecommender {
-    List<Integer> recommend(String solvedProblemNumbers);
+    List<Long> recommend(String solvedProblemNumbers);
 }
