+++ server/src/main/java/com/example/server/student/problem/dto/ItemResponse.java
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ItemResponse {
 
-    int problemId;
+    Long problemId;
     String titleKo;
 }
